requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-timezone-field==7.1
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
eventlet==0.39.0
Flask==3.1.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
MouseInfo==0.1.3
packaging==24.2
phonenumbers==9.0.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg==3.2.4
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wikipedia==1.4.0
